requirements.txt
google-cloud-speech
google-cloud-storage
pyaudio
